requirements.txt
Flask==2.3.3
Flask-Session==0.4.0
requests==2.31.0
websocket-client==1.6.1
python-dotenv==1.0.0